requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.1
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.5.1
apache-airflow-providers-common-io==1.5.1
apache-airflow-providers-common-sql==1.24.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.3
apache-airflow-providers-http==5.2.1
apache-airflow-providers-imap==3.8.3
apache-airflow-providers-smtp==2.0.1
apache-airflow-providers-sqlite==4.0.1
apispec==6.8.1
appnope==0.1.4
argcomplete==3.6.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.1
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
gcloud==0.18.3
google-api-core==2.24.2
google-auth==2.38.0
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-re2==1.1.20240702
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
ipykernel==6.29.5
ipython==8.34.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy-object-proxy==1.10.0
limits==4.4.1
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
multidict==6.1.0
narwhals==1.30.0
nest-asyncio==1.6.0
numpy==2.2.3
oauth2client==4.1.3
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
# pyspark==3.5.5
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.23.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setproctitle==1.3.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
time-machine==2.16.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
xgboost==3.0.0
yarl==1.18.3
zipp==3.21.0
